feat: Документация, рефакторинг, Readme, тестирование на мелкие недочеты. (issue-11)
