+++ src/shared/api/types.ts
@@ -5,4 +5,5 @@ export type DepartmentDto =
   | "Логистика"
   | "Сбыт";
 
+// Тип фильтра который можно примешивать в параметру фильтра managerId_ne
 export type FilterType = "lt" | "lte" | "gt" | "gte" | "eq" | "ne" | "in";
