+++ src/shared/constants/department.ts
@@ -1,5 +1,8 @@
 import type { Department } from "shared/types";
 
+// Добавлены Enums для удобства использования в коде дабы избежать ошибок при написании строк, и выбирать значения из списка
+// Сделано через const as const потому что Enum в typescript нельзя расширять, объединять и там нет сравнения по строке
+
 export const DEPARTMENT_ENUM = {
   RECONNAISSANCE: "Разведка",
   EXTRACTION: "Добыча",
