+++ src/shared/constants/urls.ts
@@ -1,1 +1,3 @@
-export const API_URL = "http://localhost:3001";
+// Базовый url для api запросов
+
+export const API_URL = process.env.REACT_APP_API_URL || "";
