+++ src/shared/lib/formatters.ts
@@ -1,5 +1,7 @@
 import { CURRENT_LOCALE } from "shared/constants";
+// Утилиты форматирования
 
+// Форматирование бюджета
 export const formatBudget = (
   value: number | string,
   currency: string = "RUB",
@@ -13,6 +15,7 @@ export const formatBudget = (
   }).format(value);
 };
 
+// вывод даты в нужном формате для локали
 export const formatDate = (value: string | Date): string => {
   if (!value) return "—";
   return new Date(value).toLocaleDateString(CURRENT_LOCALE);
